+++ src/constants/oauth.ts
@@ -1,0 +1,20 @@
+export const OAUTH = [
+  {
+    key: 'naver',
+    bgColor: '#03CF5D',
+    icon: '/assets/icons/ic_naver.svg',
+    size: 26
+  },
+  {
+    key: 'google',
+    bgColor: '#FFFFFF',
+    icon: '/assets/icons/ic_google.svg',
+    size: 32
+  },
+  {
+    key: 'kakao',
+    bgColor: '#FEE500',
+    icon: '/assets/icons/ic_kakaotalk.svg',
+    size: 38
+  }
+];
