+++ src/types/login.ts
@@ -1,0 +1,1 @@
+export type loginType = 'naver' | 'google' | 'kakao';
